last few led's (36C,06,22,exec,03,corr) now also light individualy. Still some confusion when pressing buttons: press exec lights up 3, press and hold corr lights up multiple others

+++ 50227/leds50227.c
@@ -87,9 +87,9 @@ unsigned char led_get(unsigned char at)
 void led_set(unsigned char at, unsigned char onoff)
 {
 	if (onoff)
-		_LEDs |= (1 << at);
+		_LEDs |= ((uint32_t)1 << at);
 	else
-		_LEDs &= ~(1 << at);
+		_LEDs &= ~((uint32_t)1 << at);
 	_led_update_col(at/6, (_LEDs >> (((unsigned char)(at/6)) * 6)) & 0x3F);
 	//couldn't this also be led_update(at); ?
 }
